backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.1
httpx==0.26.0
temporalio==1.4.0
openai==1.12.0
websockets==12.0
loguru==0.7.2
